Design: 결제 완료, 결제 실패 페이지 레이아웃 생성
