+++ src/main/resources/templates/thymeleaf/Member/success.html
@@ -5,14 +5,36 @@
 
 <body>
 	<div th:replace="~{nav/nav::nav}"></div>
-	
-	<div class="container">
-	
-    <h2>결제 성공</h2>
-    <p id="paymentKey">환불키</p>
-    <p id="orderId">주문번호</p>
-    <p id="amount">결제 금액</p>
+
+	<div class="container p-5 text-center">
+		<div class=" my-5 w-75 pt-5 m-auto row">
+			<div class="col-12 col-md-7 m-auto p-5 shadow-s rounded-5">
+				<div class="row gap-4">
+					<div>
+						<img src="../img/mypage/check.gif" alt="" width="48" height="48" />
+						<span class="h5 d-block m-3">결제 성공!</span> 
+					</div>
+						<img
+							src="../img/mypage/bg_detail_pop.png" alt=""
+							class="img-fluid w-100" />
+					<!-- <p id="paymentKey">환불키</p> -->
+					<div class="text-start px-3 px-lg-5">
+						<p>
+							DATE : <span>2024-03-09</span>
+						</p>
+						<span>결제 코스 : </span><p id="orderId"></p>
+						<span>결제 금액 : </span><p id="amount"></p>
+					</div>
+						<img
+							src="../img/mypage/bg_detail_pop.png" alt=""
+							class="img-fluid w-100" />
+					<a href="/home"
+						class="btn btn-main text-center d-block my-3 py-2 rounded-4 col-12 col-lg-5 m-auto">홈으로
+						돌아가기</a>
+				</div>
+			</div>
+		</div>
 	</div>
-	
-  </body>
+
+</body>
 </html>
