+++ src/main/resources/templates/thymeleaf/feed/feed_show.html
@@ -149,7 +149,7 @@
 							</div>
 							<div class="col flex-grow-1 p-0 f14">
 								<strong><span class="" th:text="${comment.getMember_Id()}">장창엽</span></strong>
-								<span class="ms-1 text-muted f12" th:text="${comment.getFeed_comment_updateday()}">댓글 단 시간</span>
+								<span class="ms-1 text-muted text-truncate f12" th:text="${comment.getFeed_comment_updateday()}">댓글 단 시간</span>
 								<span class="f14 d-block" th:text="${comment.getFeed_comment()}">오늘 쫌 힘드네여</span>
 							</div>
 							<div class="col-2 flex-shrink-0 d-sm-flex justify-content-end align-items-center p-0 h-100 text-end">
@@ -158,7 +158,7 @@
 										<i class="fi fi-br-heart px-2"></i>
 									</a>
 								</th:block>
-								<th:block th:if="${comment.getFeed_comment_like()}==1">
+								<th:block th:if="${comment.getFeed_comment_like()}==1">w
 									<a class="dropdown-item" th:href="@{'/feed_comment_like?num='+${feedList.getFeed_num()}+'&comment_num='+${comment.getFeed_comment_num()}+'&on=0'}">
 										<i class="fi fi-sr-heart px-2"></i>
 									</a>
